All-new API for efficiency and simplicity.
This uses a single factory, RxSharedPreferences, for creating all of the preference objects. Instances of this class register once with the preferences to listen to changes allowing individual preferences to just filter the list.

Preference objects are now unified into a single type with an 'adapter' responsible for getting and setting values of the correct type in the underlying preferences. This adapter mechanism is re-used for allowing custom types to be stored beyond just the built-in types.